refactor(pipeline): extract ingestion pipeline into modular architecture
- Extract pipeline components to polylogue/pipeline/
- Create dedicated modules for IDs, models, ingest preparation, and runner
- Update polylogue/run.py to act as a backward-compatible shim
- Implement progress callback support in runner
