+++ polylogue/pipeline/__init__.py
@@ -1,0 +1,29 @@
+"""Pipeline package for data ingestion and orchestration."""
+
+from polylogue.pipeline.ids import (
+    attachment_content_id,
+    attachment_seed,
+    conversation_content_hash,
+    conversation_id,
+    message_content_hash,
+    message_id,
+)
+from polylogue.pipeline.ingest import prepare_ingest
+from polylogue.pipeline.models import ExistingConversation, PlanResult, RunResult
+from polylogue.pipeline.runner import latest_run, plan_sources, run_sources
+
+__all__ = [
+    "PlanResult",
+    "RunResult",
+    "ExistingConversation",
+    "plan_sources",
+    "run_sources",
+    "latest_run",
+    "prepare_ingest",
+    "conversation_id",
+    "message_id",
+    "attachment_content_id",
+    "attachment_seed",
+    "conversation_content_hash",
+    "message_content_hash",
+]
